requirements.txt
Flask==2.3.3
tensorflow>=2.12.0
opencv-python>=4.7.0
numpy>=1.24.0
pillow>=10.0.0
h5py>=3.8.0
scikit-learn>=1.2.0
werkzeug>=2.3.0
requests>=2.31.0
